Monitor_LLM_Token_Usage/app/requirements.txt
fastapi
uvicorn
prometheus_client
python-dotenv
openai